fix: 리스트 아이템 분리 및 스타일 수정
